// src/app/modules/player-management/models/PlayerCaseModel.ts
export interface PlayerCaseModel {
	caseId: number;
	caseStatus: string;
	caseType: string;
	caseTypeId: number;
	communicationId: number;
	messageType: string;
	messageStatus: string;
	messageResponse: string;
	campaingName: string;
	createdByName: string;
	createdDate: Date;
	camapaignType: string;
	caseCreatedBy: string;
	caseCreatedDate: Date;
	communicationCreatedBy: string;
}